+++ index.js
@@ -32,15 +32,20 @@ class tasmotaPlatform {
           return;
         }
 
-        //debug config
+        //log config
         const enableDebugMode = device.enableDebugMode || false;
-        const debug = enableDebugMode ? log.info(`Device: ${host} ${deviceName}, did finish launching.`) : false;
+        const disableLogDeviceInfo = device.disableLogDeviceInfo || false;
+        const disableLogInfo = device.disableLogInfo || false;
+        const disableLogSuccess = device.disableLogSuccess || false;
+        const disableLogWarn = device.disableLogWarn || false;
+        const disableLogError = device.disableLogError || false;
+        const debug = enableDebugMode ? log.info(`Device: ${host} ${deviceName}, debug: Did finish launching.`) : false;
         const config = {
           ...device,
           user: 'removed',
           passwd: 'removed'
         };
-        const debug1 = enableDebugMode ? log.info(`Device: ${host} ${deviceName}, Config: ${JSON.stringify(config, null, 2)}.`) : false;
+        const debug1 = !enableDebugMode ? false : log.info(`Device: ${host} ${deviceName}, Config: ${JSON.stringify(config, null, 2)}.`);
 
         //check files exists, if not then create it
         const postFix = device.host.split('.').join('');
@@ -70,25 +75,25 @@ class tasmotaPlatform {
           const tasmotaDevice = new TasmotaDevice(api, device, miElHvac, defaultHeatingSetTemperatureFile, defaultCoolingSetTemperatureFile);
           tasmotaDevice.on('publishAccessory', (accessory) => {
             api.publishExternalAccessories(PluginName, [accessory]);
-            log.success(`Device: ${host} ${deviceName}, Published as external accessory.`);
+            const emitLog = disableLogSuccess ? false : log.success(`Device: ${host} ${deviceName}, Published as external accessory.`);
           })
             .on('devInfo', (devInfo) => {
-              log.info(devInfo);
+              const emitLog = disableLogDeviceInfo ? false : log.info(devInfo);
             })
-            .on('success', (message) => {
-              log.success(`Device: ${host} ${deviceName}, ${message}.`);
+            .on('success', (success) => {
+              const emitLog = disableLogSuccess ? false : log.success(`Device: ${host} ${deviceName}, ${success}.`);
             })
-            .on('message', (message) => {
-              log.info(`Device: ${host} ${deviceName}, ${message}.`);
+            .on('info', (info) => {
+              const emitLog = disableLogInfo ? false : log.info(`Device: ${host} ${deviceName}, ${info}.`);
             })
             .on('debug', (debug) => {
-              log.info(`Device: ${host} ${deviceName}, debug: ${debug}.`);
+              const emitLog = !enableDebugMode ? false : log.info(`Device: ${host} ${deviceName}, debug: ${debug}.`);
             })
             .on('warn', (warn) => {
-              log.warn(`Device: ${host} ${deviceName}: ${warn}.`);
+              const emitLog = disableLogWarn ? false : log.warn(`Device: ${host} ${deviceName}, ${warn}.`);
             })
-            .on('error', async (error) => {
-              log.error(`Device: ${host} ${deviceName}, ${error}.`);
+            .on('error', (error) => {
+              const emitLog = disableLogError ? false : log.error(`Device: ${host} ${deviceName}, ${error}.`);
             });
 
           //create impulse generator
@@ -98,16 +103,16 @@ class tasmotaPlatform {
               const startDone = await tasmotaDevice.start();
               const stopImpulseGenerator = startDone ? await impulseGenerator.stop() : false;
             } catch (error) {
-              log.error(`Device: ${host} ${deviceName}, ${error}, trying again.`);
+              const emitLog = disableLogError ? false : log.error(`Device: ${host} ${deviceName}, ${error}, trying again.`);
             };
           }).on('state', (state) => {
-            const debug = enableDebugMode ? state ? log.info(`Device: ${host} ${deviceName}, Start impulse generator started.`) : log.info(`Device: ${host} ${deviceName}, Start impulse generator stopped.`) : false;
+            const emitLog = !enableDebugMode ? false : state ? log.info(`Device: ${host} ${deviceName}, Start impulse generator started.`) : log.info(`Device: ${host} ${deviceName}, Start impulse generator stopped.`);
           });
 
           //start impulse generator
           await impulseGenerator.start([{ name: 'start', sampling: 45000 }]);
         } catch (error) {
-          log.error(`Device: ${host} ${deviceName}, Did finish launch error: ${error}.`);
+          throw new Error(`Device: ${host} ${deviceName}, Did finish launching error: ${error}.`);
         }
       };
     });