+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - (31.01.2025)
+
+## Changes
+
+- added possibility to disable/enable log success, info, warn, error
+- bump dependencies
+- config schema updated
+- redme updated
+- cleanup
+
 ## [1.1.8] - (05.01.2025)
 
 ## Changes